+++ src/main/java/com/github/alexisjehan/mavencheck/Application.java
@@ -84,6 +84,12 @@ public final class Application {
 	private static final String COMMAND_NAME = "mvnchk";
 
 	/**
+	 * <p>Maximum depth option long name.</p>
+	 * @since 1.1.0
+	 */
+	private static final String OPTION_MAX_DEPTH = "max-depth";
+
+	/**
 	 * <p>Help option long name.</p>
 	 * @since 1.0.0
 	 */
@@ -120,12 +126,24 @@ public final class Application {
 	private static final Path DEFAULT_PATH = Path.of(".");
 
 	/**
+	 * <p>Default maximum depth.</p>
+	 * @since 1.1.0
+	 */
+	private static final int DEFAULT_MAX_DEPTH = Integer.MAX_VALUE;
+
+	/**
 	 * <p>Options.</p>
 	 * @since 1.0.0
 	 */
 	private static final Options options = new Options();
 
 	static {
+		options.addOption(
+				"d",
+				OPTION_MAX_DEPTH,
+				true,
+				"Maximum depth of subdirectories to find build files (a non-negative integer)"
+		);
 		options.addOption(
 				"h",
 				OPTION_HELP,
@@ -228,6 +246,9 @@ public final class Application {
 						1 == arguments.length
 								? Path.of(arguments[0])
 								: DEFAULT_PATH,
+						commandLine.hasOption(OPTION_MAX_DEPTH)
+								? Integer.parseUnsignedInt(commandLine.getOptionValue(OPTION_MAX_DEPTH))
+								: DEFAULT_MAX_DEPTH,
 						commandLine.hasOption(OPTION_IGNORE_SNAPSHOTS),
 						commandLine.hasOption(OPTION_SHORT)
 				);
@@ -244,12 +265,35 @@ public final class Application {
 	 * @param short0 {@code true} if only build files with at least one artifact update should be shown
 	 * @throws IOException might occur with input/output operations
 	 * @throws NullPointerException if the path is {@code null}
+	 * @deprecated since 1.1.0, use {@link #run(Path, int, boolean, boolean)} instead
 	 * @since 1.0.0
 	 */
+	@Deprecated(since = "1.1.0")
 	void run(final Path path, final boolean ignoreSnapshots, final boolean short0) throws IOException {
+		run(path, DEFAULT_MAX_DEPTH, ignoreSnapshots, short0);
+	}
+
+	/**
+	 * <p>Run the program.</p>
+	 * @param path a path
+	 * @param maxDepth a maximum depth
+	 * @param ignoreSnapshots {@code true} if build file artifacts with a snapshot version should be ignored
+	 * @param short0 {@code true} if only build files with at least one artifact update should be shown
+	 * @throws IOException might occur with input/output operations
+	 * @throws NullPointerException if the path is {@code null}
+	 * @throws IllegalArgumentException if the maximum depth is lower than {@code 0}
+	 * @since 1.1.0
+	 */
+	void run(
+			final Path path,
+			final int maxDepth,
+			final boolean ignoreSnapshots,
+			final boolean short0
+	) throws IOException {
 		Ensure.notNull("path", path);
+		Ensure.greaterThanOrEqualTo("maxDepth", maxDepth, 0);
 		final var service = createService();
-		final var buildFiles = service.findBuildFiles(path);
+		final var buildFiles = service.findBuildFiles(path, maxDepth);
 		if (buildFiles.isEmpty()) {
 			outputStream.println("No build file found");
 			return;